% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambridge_bumps_results.R
\docType{data}
\name{cambridge_bumps_results}
\alias{cambridge_bumps_results}
\title{Cambridge University Bumps Results 2008-2018}
\format{A data frame with 1293 rows and 10 variables:
\describe{
  \item{Competition}{The competition- Lents or Mays}
  \item{College}{The college the crew represents}
  \item{Year}{The year of the race}
  \item{Crew}{The crew involved (e.g. NA indicates the first crew, 2 indicates the second crew...)}
  \item{Gender}{The gender - either M (male) of F (female)}
  \item{StartPos}{The starting position of that crew in the competition}
  \item{Day1}{Results on day 1 of the compeition (1 indicates a bump, -1 indicates being bumped, NA = no change)}
  \item{Day2}{Results on day 1 of the compeition (1 indicates a bump, -1 indicates being bumped, NA = no change)}
  \item{Day3}{Results on day 1 of the compeition (1 indicates a bump, -1 indicates being bumped, NA = no change)}
  \item{Day4}{Results on day 1 of the compeition (1 indicates a bump, -1 indicates being bumped, NA = no change)}
}}
\usage{
cambridge_bumps_results
}
\description{
Full results for Cambridge Bumps results from 2008-2018
}
\keyword{datasets}
