% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_data.R
\docType{data}
\name{college_data}
\alias{college_data}
\title{Dataset containing information on Cambridge college boat clubs}
\format{A data frame with 34 rows and 4 variables:
\describe{
  \item{college}{The full name of the college}
  \item{short_name}{The abbreviated name of the college}
  \item{club_code}{The code for the colleges boats}
  \item{club_name}{The abbreviated name of the college boat club}
}}
\usage{
college_data
}
\description{
Contains information on names,codes and colours for Cambridge college boat clubs
}
\keyword{datasets}
